pyproject.toml
[tool.poetry]
name = "coruja"
version = "0.1.0"
description = ""
authors = ["Ulian <falivith@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sounddevice = "^0.5.2"
scipy = "^1.15.3"
keyboard = "^0.13.5"
openai-whisper = "^20240930"
requests = "^2.32.3"
fastapi = "^0.115.12"
pydantic = "^2.11.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
